+++ src/Factories/Http/ServerRequestFactory.php
@@ -44,15 +44,15 @@ class ServerRequestFactory extends AbstractFactory implements ServerRequestFacto
     # Copied from Guzzles ::fromGlobals(), but we need to support the $server array as
     # paramter, so we use that instead of the $_SERVER array guzzle uses by default
 
-    $method = isset($server['REQUEST_METHOD']) ? $server['REQUEST_METHOD'] : 'GET';
+    $method = isset($_SERVER['REQUEST_METHOD']) ? $_SERVER['REQUEST_METHOD'] : 'GET';
     $headers = \function_exists('getallheaders') ? \getallheaders() : [];
     $uri = $uri;
     $body = new LazyOpenStream('php://input', 'r+');
-    $protocol = isset($server['SERVER_PROTOCOL']) ? \str_replace('HTTP/', '', $server['SERVER_PROTOCOL']) : '1.1';
+    $protocol = isset($_SERVER['SERVER_PROTOCOL']) ? \str_replace('HTTP/', '', $_SERVER['SERVER_PROTOCOL']) : '1.1';
 
-    $serverRequest = new ServerRequest($method, $uri, $headers, $body, $protocol, $server);
+    $serverRequest = new ServerRequest($method, $uri, $headers, $body, $protocol, $_SERVER);
 
-    \logger()->debug('Created PSR-7 ServerRequest("'.$method.'","'.$url.'") (Guzzle)');
+    \logger()->debug('Created PSR-7 ServerRequest("'.$method.'","'.$uri.'") (Guzzle)');
 
     return $serverRequest
         ->withCookieParams($_COOKIE)